requirements.txt
Django == 4.2
django-filter == 24.2
djangorestframework == 3.15.2
Markdown == 3.6
Pygments == 2.18.0
drf-yasg == 1.21.7
httpie == 3.2.2
Pillow==10.1.0
